Move `NexusMods.Media` into `NexusMods.UI.Sdk`
