--- src/NexusMods.Media/NexusMods.Media.csproj
@@ -1,10 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-    <ItemGroup>
-        <ProjectReference Include="..\NexusMods.Abstractions.Resources\NexusMods.Abstractions.Resources.csproj" />
-        <ProjectReference Include="..\NexusMods.DataModel\NexusMods.DataModel.csproj" />
-    </ItemGroup>
-    <ItemGroup>
-        <PackageReference Include="Microsoft.Extensions.Http.Resilience" />
-        <PackageReference Include="QoiSharp" />
-    </ItemGroup>
-</Project>
