+++ src/NexusMods.UI.Sdk/Resources/ImageDecoder.cs
@@ -1,7 +1,7 @@
-using NexusMods.Abstractions.Resources;
+using NexusMods.Sdk.Resources;
 using SkiaSharp;
 
-namespace NexusMods.Media;
+namespace NexusMods.UI.Sdk.Resources;
 
 public sealed class ImageDecoder<TResourceIdentifier> : ANestedResourceLoader<TResourceIdentifier, SKBitmap, byte[]>
     where TResourceIdentifier : notnull
