+++ src/NexusMods.UI.Sdk/Resources/ImageEncoder.cs
@@ -1,7 +1,7 @@
-using NexusMods.Abstractions.Resources;
+using NexusMods.Sdk.Resources;
 using SkiaSharp;
 
-namespace NexusMods.Media;
+namespace NexusMods.UI.Sdk.Resources;
 
 public sealed class ImageEncoder<TResourceIdentifier> : ANestedResourceLoader<TResourceIdentifier, byte[], SKBitmap>
     where TResourceIdentifier : notnull
