+++ tests/NexusMods.UI.Tests/NexusMods.UI.Tests.csproj
@@ -6,7 +6,6 @@
         <ProjectReference Include="..\..\src\NexusMods.Games.RedEngine\NexusMods.Games.RedEngine.csproj" />
         <ProjectReference Include="..\..\src\NexusMods.App.UI\NexusMods.App.UI.csproj" />
         <ProjectReference Include="..\..\src\NexusMods.App\NexusMods.App.csproj" />
-        <ProjectReference Include="..\..\src\NexusMods.Media\NexusMods.Media.csproj" />
         <ProjectReference Include="..\NexusMods.StandardGameLocators.TestHelpers\NexusMods.StandardGameLocators.TestHelpers.csproj" />
     </ItemGroup>
